rkt: move download code from cas remote to rkt fetch.
As the download code is not really related to the cas remote, this patch
factors it out from cas to rkt fetch.
It doesn't change the existing code but just moves it (with the minimum needed
changes) leaving additional refactoring for future patches.
It's now rkt fetch that should get the remote, choose if the images needs to be
downloaded, verify the signature and write it to the cas blob store and then
update the remote.

As a side effect it avoids the stdout printing done by cas.Remote.Download
(remains the printing done by docker2aci)

This will also help avoiding tricky things and code duplication for future etag
and cache-control handling. See #319.
