+++ cas/cas_test.go
@@ -15,13 +15,10 @@
 package cas
 
 import (
-	"archive/tar"
 	"bytes"
 	"database/sql"
 	"encoding/hex"
 	"io/ioutil"
-	"net/http"
-	"net/http/httptest"
 	"os"
 	"testing"
 	"time"
@@ -46,91 +43,6 @@ func TestBlobStore(t *testing.T) {
 		"I am a manually placed object",
 	} {
 		ds.stores[blobType].Write(types.NewHashSHA512([]byte(valueStr)).String(), []byte(valueStr))
-	}
-
-	ds.Dump(false)
-}
-
-func TestDownloading(t *testing.T) {
-	dir, err := ioutil.TempDir("", tstprefix)
-	if err != nil {
-		t.Fatalf("error creating tempdir: %v", err)
-	}
-	defer os.RemoveAll(dir)
-
-	imj := `{
-			"acKind": "ImageManifest",
-			"acVersion": "0.4.0",
-			"name": "example.com/test01"
-		}`
-
-	entries := []*aci.ACIEntry{
-		// An empty file
-		{
-			Contents: "hello",
-			Header: &tar.Header{
-				Name: "rootfs/file01.txt",
-				Size: 5,
-			},
-		},
-	}
-
-	aci, err := aci.NewACI(dir, imj, entries)
-	if err != nil {
-		t.Fatalf("error creating test tar: %v", err)
-	}
-
-	// Rewind the ACI
-	if _, err := aci.Seek(0, 0); err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	body, err := ioutil.ReadAll(aci)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Write(body)
-	}))
-	defer ts.Close()
-
-	tests := []struct {
-		ACIURL string
-		SigURL string
-		body   []byte
-		hit    bool
-	}{
-		// The Blob entry isn't used
-		{ts.URL, "", body, false},
-		{ts.URL, "", body, true},
-	}
-
-	ds, err := NewStore(dir)
-	if err != nil {
-		t.Fatalf("Unexpected error: %v", err)
-	}
-
-	for _, tt := range tests {
-		_, ok, err := ds.GetRemote(tt.ACIURL)
-		if err != nil {
-			t.Fatalf("unexpected err: %v", err)
-		}
-		if tt.hit == false && ok {
-			panic("expected miss got a hit")
-		}
-		if tt.hit == true && !ok {
-			panic("expected a hit got a miss")
-		}
-		rem := NewRemote(tt.ACIURL, tt.SigURL)
-		_, aciFile, err := rem.Download(*ds, nil)
-		if err != nil {
-			t.Fatalf("error downloading aci: %v", err)
-		}
-		defer os.Remove(aciFile.Name())
-
-		_, err = rem.Store(*ds, aciFile, false)
-		if err != nil {
-			panic(err)
-		}
 	}
 
 	ds.Dump(false)
